- Updates implementation of rmc_send and rmc_recv - App-level fixes to use the new interface (rpc_service, rpc_sync) - Compilation fixes to rmcd

+++ libsonuma/sonuma_daemon.c
@@ -189,7 +189,7 @@ void rmc_send(rmc_wq_t *wq, char *lbuff_ptr, int lbuff_offset, size_t size, int 
     }
 }
 
-void rmc_recv(rmc_wq_t *wq, char* lbuff_ptr,int lbuff_offset,size_t size,int snid,unsigned sending_qp,unsigned slot_idx)
+void rmc_recv(rmc_wq_t *wq, char* lbuff_ptr,int lbuff_offset,size_t size,int snid,uint8_t sending_qp,uint8_t slot_idx)
 {
     // create WQ entry, response for arguments given to CQ
     uint8_t wq_head = wq->head;
